requirements.txt
lastpass-python
